Changed uniqueness constraints to:   Datafile: dataset, name (-location)   Dataset: investigation, name (-sample -type)   Application: facility, name, version (+facility)   Sample: investigation, name (-type)   Investigation: facility, name, visitid (-facilityCycle)   SampleType: facility, name, molecularFormula (+molecularFormula)
Made all constraint fields not nullable. This is checked by ICAT to ensure that future changes do not break this rule.

Added a string attribute: "arguments" to the job to store the program arguments that were used.
          
Relationship between FacilityCycle and Investigation removed.

Table between Investigation and Instrument added to represent many-to-many relationships between them.